backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 数据库
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Google Gemini AI
google-genai==1.0.0

# 工具库
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# 图片处理
pillow==10.4.0

# 定时任务
apscheduler==3.10.4
